Finished functionality of settings menu
There is a strange bug with the animator that does not allow me to interact with anything. I can move the states around, but cannot make transitions, and I have to right click on the "exit" state to create new states.

Additionally, there are no comments in my code, but I will add them.